requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Async Task Queue
celery==5.3.4
redis==5.0.1

# Monitoring & Performance
django-redis==5.4.0
celery-progress==0.3

# Code Quality
black==23.11.0
ruff==0.1.6

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# API Documentation
drf-spectacular==0.26.5

# WSGI Server
gunicorn==21.2.0

# Utilities
python-json-logger==2.0.7
